requirements.txt
emoji
requests
python-telegram-bot